[CBRD-23260] Separate table/descriptor files (#1909)
http://jira.cubrid.org/browse/CBRD-23260

Separate table/descriptor files. lockfree_transaction files become lockfree_transaction_table and descriptor is moved to its own file.

Also tried to clarify comments for each component of lockfree transaction implementation.
