+++ sa/CMakeLists.txt
@@ -124,7 +124,8 @@ set(BASE_SOURCES
   ${BASE_DIR}/tsc_timer.c
   ${BASE_DIR}/lock_free.c
   ${BASE_DIR}/lockfree_bitmap.cpp
-  ${BASE_DIR}/lockfree_transaction.cpp
+  ${BASE_DIR}/lockfree_transaction_descriptor.cpp
+  ${BASE_DIR}/lockfree_transaction_table.cpp
   ${BASE_DIR}/lockfree_transaction_system.cpp
   ${BASE_DIR}/fileline_location.cpp
   ${BASE_DIR}/resource_tracker.cpp
@@ -139,8 +140,9 @@ set(BASE_HEADERS
   ${BASE_DIR}/fileline_location.hpp
   ${BASE_DIR}/lockfree_bitmap.hpp
   ${BASE_DIR}/lockfree_freelist.hpp
-  ${BASE_DIR}/lockfree_transaction.hpp
   ${BASE_DIR}/lockfree_transaction_def.hpp
+  ${BASE_DIR}/lockfree_transaction_descriptor.hpp
+  ${BASE_DIR}/lockfree_transaction_table.hpp
   ${BASE_DIR}/lockfree_transaction_system.hpp
   ${BASE_DIR}/mem_block.hpp
   ${BASE_DIR}/memory_private_allocator.cpp
